Added beeping to 5 min countdown

+++ include/SpeakerControl.hpp
@@ -7,7 +7,7 @@ class SpeakerControl
 {
 public:
     SpeakerControl();
-    void beep(int seconds);
+    void beep(double seconds);
 
 private:
     // Speaker pins
