+++ src/SpeakerControl.cpp
@@ -7,11 +7,11 @@ SpeakerControl::SpeakerControl()
     pinMode(posPin, OUTPUT);
 }
 
-void SpeakerControl::beep(int seconds)
+void SpeakerControl::beep(double seconds)
 {
     analogWrite(negPin, -128);
     analogWrite(negPin, 128);
-    delay(seconds*1000); // Beep for 0.5 seconds
+    delay(seconds*1000);
     analogWrite(negPin, 0);
     analogWrite(negPin, 0);
 }
